pyproject.toml
[project]
name = "system-metrics-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-pubsub>=2.27.1",
    "psutil>=6.1.0",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]

[tool.uv.workspace]
members = ["test"]
